refactor(components): extract Header, Footer, and Main components from HomePage
