+++ src/components/common/Logo.tsx
@@ -7,7 +7,7 @@ const Logo = () => {
     return (
         <Link to="/" className="flex items-center space-x-2" aria-label="Go to about page">  
         <img src={vinylLog} alt=" " className="h-10 w-10" /> 
-        <span className="text-xl font-bold">Font Hill Records</span>
+        <h1 className="text-xl font-bold">Font Hill Records</h1>
         </Link>
       )
 }
