+++ src/components/layout/Footer.tsx
@@ -1,0 +1,15 @@
+const Footer = () => {
+    return (
+      <footer role="contentinfo" className="bg-gray-800 text-white py-8 w-full">
+        <div className="text-center">
+          <p>Footer</p>
+        </div>
+      </footer>
+    );
+  };
+  
+  export default Footer;
+
+
+
+
